% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureApp.R
\docType{data}
\name{colorblind_palette}
\alias{colorblind_palette}
\title{Figure Shiny App}
\format{
An object of class \code{character} of length 12.
}
\usage{
colorblind_palette
}
\description{
Figure Shiny App
}
\keyword{datasets}
